Переименовал файл add-view в edit-view, Использовал метод _setState при изменении состояния, Убрал placeHolder в разметке
